move the new test to datasource module
